Refactor mechanical setup to mechanical package
