+++ src/dip_coater/mechanical/mechanical_setup.py
